handle JAVA_HOME more gracefully
Previously, the conditional setting of `JAVA_HOME` would fail in certain environments resulting in the generated local CA not being added to the java trust store. I suspect this is down to the bash interpreter used. In addition to changing the interpreter, also make the output around JAVA_HOME more explicit.