add start script
Adds a start script with an option to ignore an already running instance of nginx.

This will be useful when scripting the starting of a service as it can become:

```
dev-nginx start -i
sbt run
```

This helps avoid the situation where you receive an error such as `ERR_CONNECTION_REFUSED` in the browser as you've forgotten to manually start nginx.